Move code to a separate location.
