+++ bcnz/bin/__init__.py
@@ -1,0 +1,1 @@
+from .librun import DummyObject, run_photoz_dask
